requirements.txt
bson==0.5.10
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
Faker==18.4.0
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
termcolor==2.3.0
Werkzeug==2.2.3
pandas
